feat/삼각형 생성
정삼각형이 아니라 반쪽짜리 직각삼각형 입니다. 수정 필요.
