+++ Projects/learnMetal/Application/Assets.xcassets/AppIcon.appiconset/Contents.json
