+++ Projects/learnMetal/Application/Assets.xcassets/Contents.json
